Add files via upload
Fixed some bugs related to different page numbers and similar files in your local repo

+++ web-scrapping.py
@@ -2,20 +2,16 @@ from bs4 import BeautifulSoup
 import requests  
 import os
 
-########### SCRIPT To SAVE ALL AC SUBMISSION OF YOUR CF PROFILE ##########
+###########SCRIPT To SAVE ALL AC SUBMISSION OF YOUR CF PROFILE ##########
 
 # Provide the below details for your codeforces profile
 
-cf_handle=""                    # Your cf handle
-no_of_pages=1                   # No of submission pages in your profile
+cf_handle="kkb_10"              # Your cf handle
+no_of_pages=10                  # No of submission pages in your profile
 path="D:\code\Code-Library\ "   # Path where you want to save all AC submssion in your local repo
-partymemberid="912345"          # Inspect element any submission id from your profile and get your partymemberid from there
+partymemberid="934927"          # Inspect element any submission id from your profile and get your partymemberid from there
   
-
-############# NOW THE MAGIC BEGINS - BY Kirti Kunj Bajpai ################
-
-
-
+############# NOW THE MAGIC BEGINS - BY Kirti Kunj Bajpai ###############
 
 url="https://codeforces.com/submissions/" + cf_handle + "/page/"
 
@@ -24,8 +20,8 @@ print("Scrapping the AC Solutions of "+cf_handle + "...........", end="\n\n\n")
 for i in range (1,no_of_pages+1):
     print("Scrapping Page " + str(i) + "......")
     print("***********************************",end="\n\n")
-    url =url + str(i)
-    html_content = requests.get(url).text  
+    req_url =url + str(i)
+    html_content = requests.get(req_url).text  
     
     soup = BeautifulSoup(html_content, "html5lib")  
 
@@ -58,15 +54,16 @@ for i in range (1,no_of_pages+1):
             # Storing into the local repo
 
             final_path = path + problem_name +".txt"
-            
-            file = open(final_path, "w+") 
-            file.write(code)
-            file.close()
+            req_path = path + problem_name + "." + subm_type
+            if not os.path.exists(path + problem_name + "." + subm_type): # If that file name already exists, then it doesn't add it to repo
+                file = open(final_path, "w+") 
+                file.write(code)
+                file.close()
 
-            # converting the file
-            base = os.path.splitext(final_path)[0]
-            ext="." + subm_type
-            os.rename(final_path, base + ext)
+                # converting the file
+                base = os.path.splitext(final_path)[0]
+                ext="." + subm_type
+                os.rename(final_path, base + ext)
 
             print(problem_name)
             print("Saving file at: " + path + problem_name + "." + subm_type ,end="\n\n")
